TS: DownloadForm/DBChoiceInput
