+++ src/components/DownloadForm/InputGroup/style.css
@@ -1,15 +1,27 @@
+@import "../../../styles/colors.css";
+
 .new-input-group {
+  --group-height: 2.5rem;
   display: flex;
   width: 100%;
-  height: 2.5rem;
+  height: var(--group-height);
   & > * {
-    height: 2.5rem;
-    line-height: 2.5rem;
+    height: var(--group-height);
+    line-height: var(--group-height);
     margin: 0;
+  }
+  & > input {
+    padding: 0;
+    border: 1px solid var(--colors-slightgray);
   }
 }
 .new-input-group-button button {
+  height: var(--group-height);
+  color: var(--colors-almost-white);
   background-color: var(--colors-submain);
+  border: 1px solid var(--colors-slightgray);
+  font-weight: bold;
+  padding: 0 1rem;
   &:hover {
     background-color: var(--colors-submain-darker);
   }
